+++ tests/class.global.test.mjs
@@ -8,4 +8,5 @@ test('class.global', () => {
     expect(is_subclass_of).not.toBeUndefined();
     expect(method_exists).not.toBeUndefined();
     expect(property_exists).not.toBeUndefined();
+    expect(get_object_vars).not.toBeUndefined();
 });